Перевод запросов к БД на асинхронный вид.
