+++ backend/api_fast_api/config.py
@@ -21,6 +21,9 @@ BASE_DIR = os.path.abspath(os.path.dirname(__file__))
 # Подключение к базе данных SQLite
 SQLALCHEMY_DATABASE_URL = 'sqlite:///' + os.path.join(BASE_DIR, 'db_api', 'api_data.db')
 
+# Подключение к АССИНХРОННОЙ базе данных SQLite
+ASYNC_SQLALCHEMY_DATABASE_URL = 'sqlite+aiosqlite:///' + os.path.join(BASE_DIR, 'db_api', 'async_api_data.db')
+
 # Путь к папке статических файлов static
 STATIC_FOLDER_PATH = os.path.join(BASE_DIR, 'static')
 # Путь к папке с шаблонами html(jinja2)
