+++ backend/notes_project.txt
@@ -145,3 +145,6 @@ http://127.0.0.10:8888/api_admin/change_lesson_data/{lesson_id}
 
 Маршрут получения данных урока из БД по ID
 http://127.0.0.10:8888/api_admin/get_lesson_data/{lesson_id}
+
+---------------------------------------- 2024-07-17
+Переход на ассинхронные запросы к БД
